refactor: Refactor the PictureCard component
Refactor the PictureCard component to include the ability to upload multiple images. Update the handleFileUpload function to append the uploaded image to the imageSources array. Update the PictureCard component to use the initialImageSources prop instead of the deprecated imageSources prop.

Update ProductsModal component

Update the ProductsModal component in the seller directory to use the initialImageSources prop instead of the deprecated imageSources prop.

Update ActivityModal component

Update the ActivityModal component in the advertiser directory to use the initialImageSources prop instead of the deprecated imageSources prop.

Update HistoricalPlacesModal component

Update the HistoricalPlacesModal component in the tourism-governor directory to use the initialImageSources prop instead of the deprecated imageSources prop.

Update ItineraryModal component

Update the ItineraryModal component in the tour-guide directory to use the initialImageSources prop instead of the deprecated imageSources prop.