Infra/prepare for production (#329)
* Add build command to package.json files

* Modify tsconfig.json files for proper building

* fix: Resolve type error in build process of api-gateway service

* fix: Resolve type errors in build process of booking service

* fix: Resolve type errors in build process of entertainment service

* Configure docker for production

* Update package.json and tsconfig.json

* Update services URLs to use prod env variables

* fix: Allow user service build to emit JS

* Use env variables for services ULRs with vite

* Allow all domains to bypass CORS for ease of deoplyments

* fix: Unable to read jwt cookies in prod environment

Changed jwt cookie props to be httpOnly, secure, sameSite=None so that it can be set and sent properly to the the backend through the cross-site request (backend origin is different from the frontend)

Also the frontend can't read cookies that have domain attribute different from the frontend host domain, hence we added a new endpoint that returns the decoded token information after making sure the token is valid. This should work as replacement of the approach where the cookie value was read and decoded on the frontend

* chore: Update compose files

* fix: Minor build errors

* chore: Format OTP email message

* fix: Unable to upload to firebase in cloud run production

Due to problem accessing firebase bucket and upload to it from cloud run, we used Supabase instead which offer more storage for free and working well with cloud run

* chore: Format email message, remove comments, add logs

* infra: Update .env.example files to be more descriptive

* chore: Delete dummy test files

* chore: Use default port and mongo URIs

* infra: Prevent docker containers from installing deps on each run

* chore: Update names of Supabase env variables

* fix: Update exchange rates API integration and add backup rates (#326)

* fix: Update exchange rates API integration and add backup rates

* Implement error handling in exchange rates fetching

Added error handling for fetching exchange rates and fallback to backup rates.

* Update error logging for exchange rates fetch

* chore: Update .env.example files to be more informative

* fix: Environment variables are not loaded properly in dev containers

The problem was that some ts files didn't have `dotenv.confg()` which resulted the env variables to not be loaded

* docs: Update .env.example files

* docs: Update .env.example

* fix: TypeScript build error

* feat: Prepare the frontend UI for the puplic

- Fix general bugs and ensure UI consistency
- Make the UI responsive for mobile and large tablets view
- Add visible loading state

* chore: Remove scrapper api keys from the .env.example

Since we temporary marked the stays section as coming soon, there is no use for the scrapping functionality and hence we should remove the scrapping api keys from the required env variables in the .env.example file to avoid confusion

---------

Co-authored-by: Abdelrahman M.Talaat <abdelrahmanmt96@gmail.com>